requirements.txt
streamlit>=1.36
gspread>=6.0.0
google-auth>=2.0.0
pandas>=2.0.0
